+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import "./App.css";
 import "./components/defaultCss/style.css";
 import Menu from "./components/menu";
-import Header from "./components/header/header";
+import Header from "./components/header";
 import Sobre from "./components/sobre/sobre";
 import Portfolio from "./components/portfolio/";
 import Contato from "./components/contato/contato";
